Update ibamr notebooks and remove bash scripts
The Jupyter notebooks now contain the branch and tag of the version of IBAMR
(forked repository under my account) used to run the simulations.

The command-line to generate the executable of the example and to run the case
is also written in each notebook.
Therefore, I removed the bash scripts.