Clean-up scripts to generate figures of the paper
Add a YAML file to indicate the directory of the simulations to use to generate
the figures.
Remove log files that are not really necessary.

The script `plotVorticity.py` creates 3 files:
* `ibamr_vorticityRe2000_tractionFreeOutlet.png`: vorticity field after about
61 time-units of flow-simulation when using a traction-free outlet boundary
condition;
* `ibamr_vorticityRe2000_stabilizedOutlet.png`: vorticity field after about
61 time-units of flow-simulation when using a traction-free condition with
stabilization at the outlet;
* `ibamr_vorticityRe2000.pdf`: the two previous files concatenated into a .pdf
(this is the figure that goes in the manuscript).

The script `plotForceCoefficientsVsAoA.py` creates 2 files:
* `ibamr_forceCoefficientsVsAoA.png`: time-averaged force coefficients;
* `ibamr_forceCoefficientsVsAoA.pdf`: the .pdf version (that should go in the
manuscript).
In those figures, I added the time-averaged force coefficients from the
simulation using a body discretization and a smaller CFL constraint.

The script `plotForceCoefficientsRe2000.py` creates 3 files:
* `ibamr_vorticityRe2000AoA30.png`: instantaneous force coefficients at Re=2000
and AoA=30deg;
* `ibamr_vorticityRe2000AoA35.png`: instantaneous force coefficients at Re=2000
and AoA=35deg (I am now display the force coefficients obtained with a
simulation using a smaller CFL constraint);
* `ibamr_vorticityRe2000.pdf`: contatenation of the two above .png files into a
single .pdf file (this figure should go in the manuscript).