Clean-up scripts to generate the PetIBM figures
Add a YAML file that contains the directory of the PetIBM simulations
to use to generate the figures.
Remove unecessary log files and bash scripts.

There are three Python scripts, each one generates a figure of the manuscript:
* `plotForceCoefficientsVsAoA.py`: plots the time-averaged force coefficients;
* `plotForceCoefficientsRe2000AoA35.py`: plots the instantaneous force
coefficients from the simulation with displaced markers (comparing to Krishnan
et al., 2014) and from the simulation with the original markers (comparing to
the results obtained with the displaced ones);
* `plotVorticity.py`: creates the plots of the vorticity field at various
time-units of the simulation at Re=2000 and AoA=35deg, that uses the original
markers.

The script `plotForceCoefficientsVsAoA.py` creates 2 files:
* `petibm011_forceCoefficientsVsAoA.png` (to be used in markdown);
* `petibm011_forceCoefficientsVsAoA.pdf` (to be used in the manuscript).

The script `plotForceCoefficientsRe2000AoA35.py` creates 3 files:
* `petibm011_forceCoefficientsRe2000AoA35CompareKrishnan.png`: compares the
instantaneous force coefficients from the simulation using displaced markers to
those reported in Krishnan et al. (2014);
* `petibm011_forceCoefficientsRe2000AoA35CompareMarkers.png`: compares the
instantaneous force coefficients between the simulation using displaced markers
and the one using the original markers of Krishnan et al. (2014);
* `petibm011_forceCoefficientsRe2000AoA35.pdf`: contanetation of the 2 .png
files described above (the file should used in the manuscript).

The script `plotVorticity.py` creates 5 files:
* `petibm011_vorticity475000Re2000AoA35.png`: vorticity field after 19
time-units of flow-simulation;
* `petibm011_vorticity130000Re2000AoA35.png`: vorticity field after 52
time-units of flow-simulation;
* `petibm011_vorticity132500Re2000AoA35.png`: vorticity field after 53
time-units of flow-simulation;
* `petibm011_vorticity160000Re2000AoA35.png`: vorticity field after 64
time-units of flow-simulation;
* `petibm011_vorticityRe2000AoA35.pdf`: concatenation of the 4 files described
above into a single .pdf file (the one that should be used in the manuscript).