Clean-up script to generate the cuIBM figure
The following directory contains a YAML file that lists the directory of each
cuIBM simulation whose numerical solution is used to create the figure.

I removed the uncesseray log files and bash scripts.

The Python script creates 5 files:
* `cuibm-current-cusp051_forceCoefficientsRe1000AoA35.png`: instantaneous
force coefficients for a simulation at Re=1000 and AoA=35deg using the current
version of cuIBM and the library CUSP-0.5.1;
* `cuibm-current-cusp051_forceCoefficientsRe2000AoA30.png`: same than the one
above, except with Re=2000 and AoA=30deg;
* `cuibm-revision86-cusp040_forceCoefficientsRe2000AoA35.png`: instantaneous
force coefficients at Re=2000 and AoA=35deg using an older version of cuIBM
(revision86) and the library CUSP-0.4.0;
* `cuibm-current-revision86_forceCoefficientsRe2000AoA35.png`: comparing the
instantaneous force coefficients at Re=2000 and AoA=35deg using (1) a current
version of cuIBM with CUSP-0.5.1, (2) a current version of cuIBM with
CUSP-0.4.0, and (3) an older version of cuIBM with CUSP-0.4.0;
* `cuibm_forceCoefficients.pdf`: 2x2 assembly of the 4 figures described above
into a single .pdf file (that should be included in the manuscript).