Add second vorticity field and smaller CFL
With IBAMR, we used a boundary stabilization at the outlet to push the flow
outside the domain and avoid back propagation of vorticity.
The top part Figure 5 displays the vorticity field in the absence of boundary
stabilization.
The bottom part of Figure 5 now shows the improvement procured by this
stabilization.

I also added to Figure 6 the time-averaged force coefficients obtained for a
simulation using a smaller CFL constraint.
I added to Figure 7 the instantaneous force coefficients obtained for a
simulation with a smaller CFL constraint.